eth, internal: move package dao to internal, keep The DAO state
